+++ packages/client/src/constant/common.ts
@@ -1,4 +1,4 @@
 export const THEME_MODE_LOCAL_STORAGE_KEY = '__HELLO__THEME_MODE_KEY';
 export const THEME_MODE_ATTR = 'theme-mode';
 
-export const TOKEN_LOCAL_STORAGE_KEY = '__HELLO__TOKEN_KEY';
+export const ACCESS_TOKEN_LOCAL_STORAGE_KEY = '__HELLO__ACCESS_TOKEN_KEY';