+++ packages/server/src/guard/auth.guard.ts
@@ -1,11 +1,14 @@
-import { CanActivate, ExecutionContext } from '@nestjs/common';
+import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
 import { Observable } from 'rxjs';
 
+@Injectable()
 export class AuthGuard implements CanActivate {
   canActivate(context: ExecutionContext): boolean | Promise<boolean> | Observable<boolean> {
-    console.log('==guard==');
     const req = context.switchToHttp().getRequest();
-    const token = context.switchToRpc().getData();
+    const token = context.switchToRpc().getData()?.headers?.authorization;
+
+    console.log('==guard==', token);
 
     return true;
   }