+++ packages/server/src/modules/auth/guards/jwtAuth.guard.ts
@@ -5,7 +5,7 @@ import { Observable } from 'rxjs';
 @Injectable()
 export class JwtAuthGuard extends AuthGuard('jwt') {
   canActivate(context: ExecutionContext): boolean | Promise<boolean> | Observable<boolean> {
-    console.log('ctx', context);
+    console.log('ctx', super.canActivate(context));
     return super.canActivate(context);
   }
 }